make +afew feature, fix +notmuch/update buf kill
Using `compile` approach similar to `doom sync`. It runs in a buffer and
deletes the buffer when complete.

`afew` is and should be optional. Putting it behind a feature flag so
that if you don't have afew install, it won't be run. It's easily
available for those that want it, and easy to ignore for everyone else.

added documentation.